[Xamarin.Android.Build.Tasks] fixes for case-insensitive FS (#3463)
A simple C# code change could break our build.  Consider:

	class Foo : Java.Lang.Object { }

Changed it to (renaming `Foo` to `fOO`; note: capital-O, not zero):

	class fOO : Java.Lang.Object { }

`javac` will fail on Windows (and likely macOS if you have a case
insensitive file system) with:

	obj\Debug\android\src\md54a1bd985b0effc3cb8b7fa6ec8bb4028\Foo.java(4,8):  error: class fOO is public, should be declared in a file named fOO.java
	public class fOO

The only way to get the project working again is to do a `Clean` or
`Rebuild`.

This C# snippet illustrates the issue further:

	File.WriteAllText("Foo.txt", "Foo");    // or File.Copy()...
	File.WriteAllText("fOO.txt", "fOO");    // or File.Copy()...
	foreach (var file in Directory.GetFiles (".", "*.txt")) {
	  Console.WriteLine(file + " -> " + File.ReadAllText (file));
	}

The output is:

	.\Foo.txt -> fOO

i.e. even though we wrote to the file `fOO.txt`,
`Directory.GetFiles()` only returns the *first* filename casing used.

The only solution is to put a `File.Delete()` prior to the second file
write -- there is no other way to sanely change the casing on a file!

I added a call to `File.Delete()` to many of the
`Files.Copy*IfChanged()` methods, which appears to solve the problem.
I wrote a small battery of tests to be sure things are working.  One
thing to note is that `File.Delete()` does not throw if the file does
not exist, and so we don't need a `File.Exists()` check.  There are
tests covering this case as well.

We have been trying to *reduce* disk I/O, and so this is going to be
a slight performance hit.  My thought is we could remove any calls to
`File.SetLastAccessTimeUtc()` as they seem duplicative of
`File.SetLastWriteTimeUtc()`.

After the fixes and removal of `File.SetLastAccessTimeUtc`(), the
performance seems to be roughly the same if we look at two MSBuild
tasks that use `Files.Copy*IfChanged()` a lot.  Here is three runs of
a "fresh" build of the Xamarin.Forms project in this repo:

	Before:
	 172 ms  CopyIfChanged                              5 calls
	 151 ms  CopyIfChanged                              5 calls
	 161 ms  CopyIfChanged                              5 calls
	4603 ms  GenerateJavaStubs                          1 calls
	4368 ms  GenerateJavaStubs                          1 calls
	4529 ms  GenerateJavaStubs                          1 calls

	After:
	 157 ms  CopyIfChanged                              5 calls
	 180 ms  CopyIfChanged                              5 calls
	 150 ms  CopyIfChanged                              5 calls
	4241 ms  GenerateJavaStubs                          1 calls
	4496 ms  GenerateJavaStubs                          1 calls
	4319 ms  GenerateJavaStubs                          1 calls

Note that the `<CopyIfChanged/>` calls `Files.CopyIfChanged()` for
each input file internally, and the task itself was called 5 times
here.  `<GenerateJavaStubs/>` calls `Files.CopyIfStringChanged()` for
each Java stub.

Other changes:

  * I removed all usage of `File.SetLastAccessTimeUtc()` -- they
    seemed to be duplicate everywhere.
  * Fixed a bug I found, where if the destination file is readonly,
    `Files.CopyIfChanged()` failed...
  * Removed a comment in `Files.cs`