[Xamarin.Android.Build.Tasks] JAVAC0000 error code (#3445)
For tracking "deletebinobj" progress, we wanted to make sure we have
proper error codes for any failure that would be symptomatic of a
"deletebinobj" problem.

`javac` failing is one place that developers can't easily cause a
build failure without using `@(AndroidJavaSource)`.  If we get a
`JAVAC0000`, likely something went wrong with the Xamarin.Android
build system.

I also fixed up the test that was checking these error messages: it
was using `StringAssertEx.Contains()` that calls `Assert.Pass()`:
some of the assertions at the end of the test were skipped.