[Xamarin.Android.Build.Tasks] Validate the `android:versionCode`
Fixes #2023

The `versionCode` in the android manifest should be
an `int` value. It should fall within the range of
`0` to `2100000000`.

This commit adds some checks which will emit MSBuild errors
if we encounter an invalid `versionCode`. It also
adds some unit tests to test for decimal values and
the boundary conditions.