[Xamarin.Android.Build.Tasks] improve MAX_PATH errors (#3492)
Fixes: https://github.com/xamarin/xamarin-android/issues/2848
Context: https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file#maximum-path-length-limitation
Context: https://blogs.msdn.microsoft.com/jeremykuhne/2016/07/30/net-4-6-2-and-long-paths-on-windows-10/

A #deletebinobj problem has surfaced related to `MAX_PATH`.  We have
a discrepancy of path length between:

  * `obj\Debug\android\src` - where we generate `.java` source code
  * `obj\Debug\android\bin\classes` - `javac` compiles `.class` files

If you have a project on the edge, where `android\src` does *not* have
a long path, but `android\bin\classes` *does*, we get the issue here:

	error MSB3231: Unable to remove directory "obj\Debug\90\android\bin\classes".
	    Could not find a part of the path 'AccessibilityManagerCompat_AccessibilityStateChangeListenerImplementor.class'.

`javac` was able to create this file that is over `MAX_PATH`, and
unfortunately, .NET is throwing a `DirectoryNotFoundException` here.

An example:

	var dir = @"C:\src\xamarin-android\bin\TestDebug\temp\OnTheEdgeNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNNN\obj\Debug\android\bin\classes";
	Directory.Delete(dir, recursive: true);

This code throws this when there is a file inside the directory over
`MAX_PATH`:

	System.IO.DirectoryNotFoundException: 'Could not find a part of the path 'AccessibilityManagerCompat_AccessibilityStateChangeListenerImplementor.class'.'

A solution is to prefix `\\?\` in the `Directory.Delete()` call as
long as you are already using a full path:

	Directory.Delete(@"\\?\" + dir, recursive: true);

As long as you are running on .NET 4.6.2 or higher, .NET will pass
the path directly to the Windows APIs without modification.  This
works even without the `Local Group Policy` setting to enable
long paths.

However, we should use `\\?\` sparingly:

  * Not all `System.IO` and .NET APIs support `\\?\`-prefixed paths
  * You have to already know that the path you are using is a full
    path -- `Path.GetFullPath()` will not expand a long path!

So I think we can safely prefix `\\?\` in some cases, such as:

  * We hit a `DirectoryNotFoundException` or `PathTooLongException`
  * On Windows, we can safely "try again" with `\\?\`

We can use this in `<RemoveDirFixed/>`, and it is able to delete
directories containing long paths.

~~ Other changes ~~

  * I refactored `<RemoveDirFixed/>` a bit, to make things cleaner.
  * I added a `XA5301` error code for `MAX_PATH` issues.
  * I added several new tests around `MAX_PATH`.