[tests] AndroidUpdateResourcesTest timing should exclude Restore
Context: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=2942448&view=ms.vss-test-web.build-test-results-tab&runId=7928872&resultId=100035&paneView=debug

We have a specific test that times a design-time build and asserts
that it took less than a period of time.

It appears that in once instance the test taking too long, the
`Restore` target is what took a long time:

    10765 ms  Restore                                    1 calls

We should explicitly NuGet restore *first*, and not include that in
the timing.

I think this will also allow us to just check that things are under
10s and not check `IsRunningOnHostedAzureAgent`.