[Xamarin.Android.Build.Tasks] XA5302 warning for parallel builds (#3501)
Context: https://github.com/xamarin/xamarin-android/issues/3494#issuecomment-521256360

In the past, we have seen some cases where the IDE can run MSBuild on
the same project in parallel.  This causes inexplicable failures
throughout Xamarin.Android's MSBuild targets.

If we generate a build warning indicating this is happening, we can
have insight release over release if this happens in the wild.
We will also have a visible warning recorded in customer build logs.

To generate the warning:

 1. Write a zero-byte "lock file" at the beginning of a build
    * Emit an XA5302 warning if this file already exists
 2. Delete the "lock file" at the end of the current build, even if
    the build failed, was cancelled, etc.

This is simple and should work cross-platform, where a named `Mutex`
would only work on Windows.

To achieve this we can perform some MSBuild wizardry:

  * Register a new MSBuild target with `InitialTargets`.  This target
    will run before *anything* no matter what the top-level target was.
  * Pass an `IDisposable` to `RegisterTaskObject()` with a lifetime
    of the current build.  MSBuild will call `Dispose()` on the
    instance automatically at the end of the build.

We also need to use `try-catch`, so that:

  * The `<WriteLockFile/>` task never fails.
  * The `WriteLockFile.Dispose()` method never throws.

I added an assertion in one test to make sure the lock file does not
exist after a build.  I did some manually testing also, and things
seem to be working as expected.