[ci] Remove monodroid provisionator script (#3469)
We currently use the monodroid "provisionator" script to install the
JDK and select a specific version of Xcode.

However, we should no longer need to invoke the provisionator script,
as we're already installing the Corretto JDK and other dependencies
as part of `xaprepare`.  Furthermore, we should no longer have a
dependency on Xcode 9.2, so the default macOS compiler toolchain works.

As of commit 08e0ccb4 `$(AndroidToolchainDirectory)` is set to
`$(HOME)/Library/Android` when executing on macOS Azure Pipelines
hosts, and the JDK is installed into `$HOME/Library/Android/jdk`.

Export `$JAVA_HOME` on the macOS DevOps build so that it references
the JDK that `xaprepare` installs.

Additionally, instead of relying on `$JAVA_HOME` or ib JDK tools
existing in `$PATH` on macOS, use the value of `$(JavaSdkDirectory)`
to provide full paths to JDK tools.  If `$(JavaSdkDirectory)` is not
set, we set it to `$(AndroidToolchainDirectory)/jdk` which contains
the JDK we install.