[Xamarin.Android.Build.Tasks] improve incremental builds + designer
I was reviewing MSBuild logs inside the IDE when using the Android
designer and noticed something odd:

    Building target "_CompileToDalvikWithDx" completely.
    Input file "obj\Debug\android\bin\classes.zip" is newer than output file "obj\Debug\stamp\_CompileToDalvik.stamp".

It appears that editing an Android layout was causing
`_CompileToDalvikWithDx` to run again?

Then I looked at the build log from `SetupDependenciesForDesigner`:

    Building target "SetupDependenciesForDesigner" completely.
    Output file "obj\Debug\android\bin\classes\mono\MonoPackageManager.class" does not exist.

Ok, we no longer generate `MonoPackageManager.class`. It is included
as part of `java-runtime.jar`.

Looking closer, the `Inputs` and `Outputs` did not seem right for
`SetupDependenciesForDesigner`:

    Inputs="$(IntermediateOutputPath)android\src\mono\MonoPackageManager.java;$(_AndroidTypeMappingJavaToManaged);$(_AndroidTypeMappingManagedToJava)"
    Outputs="$(IntermediateOutputPath)android\bin\classes\mono\MonoPackageManager.class"

We no longer generate `$(_AndroidTypeMappingJavaToManaged)` or
`$(_AndroidTypeMappingManagedToJava)`, so we would also hit:

    Building target "SetupDependenciesForDesigner" completely.
    Input file "obj\Debug\android\typemap.jm" does not exist.

In the current state of affairs, `SetupDependenciesForDesigner` will
always run -- and trigger `_CompileToDalvikWithDx` on incremental
*regular* builds.

Two goals here would make things a lot better:

1. `SetupDependenciesForDesigner` should be skippable.
2. `SetupDependenciesForDesigner` should not trigger anything later if
   you do a regular build.

For no. 1 we can more appropriately write the `Inputs` and `Outputs` as:

    Inputs="$(MSBuildAllProjects);$(MonoPlatformJarPath);@(_JavaStubFiles);@(AndroidJavaSource)"
    Outputs="$(_AndroidStampDirectory)SetupDependenciesForDesigner.stamp"

This matches what we do for `_CompileJava`, and
`SetupDependenciesForDesigner` can manage its own stamp file.

For goal no. 2, however, we need an option for the `<Javac/>` MSBuild
task to *not* generate `classes.zip`. The designer does not use this
file, so we don't need to create it! We also don't want a regular
build later to see the timestamp on `classes.zip` change.

~~ Xamarin.ProjectTools ~~

In trying to write some tests, I actually found a bug in
Xamarin.ProjectTools:

* Call `Touch()` on a file
* The timestamp on the `.csproj` is updated -- when it did not change

To account for this, I reworked the `XamarinProject` class to save a
copy of the `ProjectResource` instance between builds.

These changes allowed me to add the needed tests in `DesignerTests`.