Fix Inputs for _ResolveLibraryProjectImports
After the fix in Xamarin.ProjectTools, the timestamp on a `.csproj`
file remains unchanged unless actual `.csproj` contents change. This
matches more closely with incremental builds in the real world.

This found a missing input on `_ResolveLibraryProjectImports`... It
operates on `@(AndroidAarLibrary)` so this item group needs to be
listed in `Inputs`.