[Jenkins] remove some duplicate test phases
As we slowly migrate to Azure DevOps, there are some duplicate phases
we are running on Jenkins that mostly produce noise, such as:

* MSBuild Tests - we commonly get failures related to NuGet
* BCL Tests - enough said...
* Networking-related issues on APK-based tests

For now we could replace the call:

    make run-all-tests

To run a subset:

    make run-performance-tests

This will keep the existing plots we have working. OSS contributors
will still have access to builds, etc.

Other changes:

* The `run-performance-tests` target should run the APK-based
  performance tests.
* We need to add `Mono.Android-Tests` to the APK performance tests.

Update RunApkTests.targets