[docs] Add ExportXamarinDocs, ImportXamarinDocs (#2789)
Add a few helper MSBuild targets to facilitate copying documents in
the xamarin-android repository to and from the corresponding files in
the MicrosoftDocs/xamarin-docs and
MicrosoftDocs/xamarin-engineering-docs-pr repositories.

For example, to update content from xamarin-android into a local
checkout of the xamarin-docs repository, run a command like:

	msbuild /t:ExportXamarinDocs /p:XamarinDocsPath=/path/to/xamarin-docs/ Xamarin.Android.sln

The `ExportXamarinDocs` target currently only copies the
`Documentation/guides/BuildProcess.md` guide.

Similarly, to update content from the xamarin-android into a local
checkout of the xamarin-engineering-docs-pr repository, run a command
like:

	msbuild /t:ExportXamarinEngineeringDocs /p:XamarinEngineeringDocsPath=/path/to/xamarin-engineering-docs-pr/ Xamarin.Android.sln

The `ExportXamarinEngineeringDocs` currently covers all the files in
the `Documentation/guides/messages` directory.  It also generates a
`TOC.md` file for the xamarin-engineering-docs-pr repo by running a
few regular expression replacements on the contents of the
`README.md` file.