[ci] Remove monodroid provisionator script (#3538)
Context: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=2968944&view=logs&s=1afc3bfe-122c-538b-e9ad-2a86c2efcfef&j=96fd57f5-f69e-53c7-3d47-f67e6cf9b93e

As of commit 08e0ccb4 `$(AndroidToolchainDirectory)` is set to
`$(HOME)/Library/Android` when executing on macOS Azure Pipelines
hosts, and the JDK is installed into `$HOME/Library/Android/jdk`.

Export `$JAVA_HOME` on the macOS DevOps build so that it references
the JDK that `xaprepare` installs.

Additionally, instead of relying on `$JAVA_HOME` or ib JDK tools
existing in `$PATH` on macOS, use the value of `$(JavaSdkDirectory)`
to provide full paths to JDK tools.  If `$(JavaSdkDirectory)` is not
set, we set it to `$(AndroidToolchainDirectory)/jdk` which contains
the JDK we install.

We currently use the monodroid `provisionator` script to install the
JDK and select a specific version of Xcode.

However, we should no longer need to invoke the `provisionator`
script, as we're already installing the Corretto JDK and other
dependencies as part of `xaprepare`.  Furthermore, we should no
longer have a dependency on Xcode 9.2, so the default macOS compiler
toolchain works.  Finally, making Xcode 9.2 the default breaks
notarization, as it is missing the required tools:

	notarize.rb:85:in `extract_guid': xcrun: error: unable to find utility "altool", not a developer tool or in PATH (NotarizationFailed)
		from notarize.rb:107:in `run'
		from notarize.rb:113:in `<main>'
	xcrun: error: unable to find utility "altool", not a developer tool or in PATH

Stop running `provisionator` and use the default Xcode.