[linker] Simplify java ifaces detection
Check whether the interface type is based on `IJavaObject` or
`IJavaPeerable` when deciding whether the iface should be marked
without further linker analysis.

@jonp suggested that idea during review of
https://github.com/xamarin/xamarin-android/pull/3525

I have tried the change with Topeka sample to make sure we don't break
the java ifaces linking again. The size stays the same. The set of
interfaces for which we return *true* in
`ShouldMarkInterfaceImplementation` is the same with the exception of
`IJavaObject` and `IJavaPeerable`. These would be preserved anyway.