[linker] Simplify java ifaces detection (#3541)
Context: https://github.com/xamarin/xamarin-android/pull/3525#discussion_r315978149

`MonoDroidMarkStep.ShouldMarkInterfaceImplementation()` needs to
determine if an interface is a *Java* interface, in which case any
types implementing the interface need to *continue* implementing the
interface; see also d762aa99.

Previously, we would decide that an interface is a Java interface if:

 1. It has a `[Register]` custom attribute, and
 2. The `[Register]` attribute takes 3 parameters, and the third
    parameter is not the empty string.

For example:

	[Register ("java/lang/Cloneable", "", "Java.Lang.ICloneableInvoker", ApiSince = 1)]
	public partial interface ICloneable : IJavaObject {
	}

However, there was no requirement that the interface implement
`IJavaObject`, meaning this declaration would also be detected as a
Java interface, even though it cannot be used as one:

	[Register ("bad/Example", "", "Bad.Example")]
	public partial interface IBadExample {
	}

(It cannot be used as a Java interface because we don't know how to
get a JNI handle from such an interface.)

Instead of checking the `[Register]` attribute, a Java interface
is an interface that implements either or both of `IJavaObject` and
`IJavaPeerable`:

	public interface ICloneable : IJavaObject {
	}

`[Register]` is still needed for Java Callable Wrapper generation,
but is no longer used by the linker to determine that a particular
interface is a *Java* interface that needs special treatment.

I have tried the change with Topeka sample to make sure we don't
break the Java interfaces with linking enabled.  The `.apk` size
stays the same.  The set of interfaces for which we return *true* in
`ShouldMarkInterfaceImplementation()` is the same except for
`IJavaObject` and `IJavaPeerable`.  These would be preserved anyway.