[build] Migrate assemble-docs target from monodroid
Fixes: https://github.com/xamarin/xamarin-android/issues/3086

Now that the [Xamarin.Android API docs][0] have been open sourced,
migrate the `assemble-docs` target and its dependencies from the
`docs.make` file in monodroid to a set of MSBuild targets in
xamarin-android.

Move `@(_FrameworkDirs)` and `@(_FrameworkDirsThatExist)` into their own
helper target so that the new targets can use them.

Add a trailing slash on the value of `$(FrameworkSrcDir)` so that the
new targets can follow the convention that the value of a directory path
property always includes a trailing slash.

Add the new targets that replace the old `docs.make` targets.  These
targets run `mdoc` to generate "msxdoc" Microsoft XML Documentation
format files that are used in the `.vsix` installer for Windows as well
as a monodoc format `MonoAndroid-lib.zip` archive and an accompanying
`.tree` file that are used in the `.pkg` installer for macOS.

Bring in android-api-docs as a submodule, set at the same commit as the
corresponding submodule in monodroid.

There are a few intentional differences compared to the original
`docs.make` targets:

  * The Microsoft XML Documentation files are output to the directory of
    the latest *stable* framework version rather than the latest *known*
    framework version.  This is useful because it means the MSBuild
    targets don't need access to the `$(FRAMEWORKS)` `make` variable.

  * The OpenTK-1.0 docs are only generated for the `mdoc assemble` step
    and not for the `mdoc export-msxdoc` step.  The `export-msxdoc`
    output isn't needed at the moment because the `.vsix` installer
    instead uses the `.xml` file that's generated when
    `Mono.Android.csproj` builds `OpenTK-1.0.csproj` for
    `$(TargetFrameworkVersion)` `v4.4`.

  * The OpenTK docs are omitted in favor of the OpenTK-1.0 docs because
    OpenTK has been deprecated and removed from Xamarin.Android.

  * The `pimpup-docs` target is omitted.  I checked the build logs for a
    recent commercial build and saw that while the build does compile
    `doc-relcontent-merger.exe` and `recipe-relcontent-generator.exe`,
    it never runs either one.  Starting in [monodroid@c83b93e2][1], the
    `$(PIMP_STAMP)` target that used to run `doc-relcontent-merger.exe`
    was no longer needed and was disabled.

  * The `docs/netdocs.zip` and `docs/mono-libs.zip` targets are also
    intentionally omitted.  Although commercial builds do still run
    those targets, the outputs are not used for anything.  For example,
    they aren't included in the `.pkg` installer for macOS or the
    `.vsix` installer for Windows.

    Based on the contents of the `push-docs` target in `docs.make` and
    the fact that that target was last edited in 2012, I believe the
    `netdocs.zip` and `mono-libs.zip` archives were for an old online
    documentation system that was retired long ago.

I verified that the `Mono.Android.xml` Microsoft XML Documentation file
generated by these new targets was identical to the version generated by
the `docs.make` targets.  Similarly, I verified that the new monodoc
`MonoAndroid-lib.zip` archive generated by these targets contained the
same list of `xml.summary.*` files as the `.zip` archive generated by
the `docs.make` targets.

[0]: https://github.com/xamarin/android-api-docs
[1]: https://github.com/xamarin/monodroid/commit/c83b93e20b0e6f80af7f8e19a6f1dc73f187e4f0