Check for integer overflows when allocating memory (#3549)
Context: https://github.com/android-ndk/ndk/issues/294
Context: https://github.com/android-ndk/ndk/issues/295
Context: https://bugs.llvm.org/show_bug.cgi?id=16404

Add a couple of functions to `Utils` taking advantage of the compiler
builtin functions to perform integer addition and multiplication
while checking for overflows.  If an overflow is detected, the
application is terminated.

Due to a bug in Android NDK's clang, the multiplication is not
performed using "open" types but rather, currently, only `size_t`
for the multiplication operands.

Using a template for them would result in a link error as the
compiler would generate code to use 128-bit integers to perform the
operation, attempting to call the `__muloti4` intrinsic function
which is usually defined in `libgcc`, `libcompiler_rt` or the
`libcompiler_rt-extras` libraries.  In the NDK case the 64-bit
targets do not contain implementation of the function in any of the 
libraries mentioned above.