[Xamarin.Android.Build.Tasks] Fix up lint output resource paths. (#3368)
Fixes: https://github.com/xamarin/xamarin-android/issues/3360

The [Android SDK `lint` tool][0] is run inside the
`$(IntermediateOutputPath)` directory.  Consequently, when `lint`
reports issues, the paths it reports point into `obj/...`, and these
paths are then reported within the Visual Studio windows.

Consequently, double-clicking on a reported issue will open the
the file within `$(IntermediateOutputPath)`.  Editing the file will
have no affect, because the edited file will be replaced with the
next build.

The result is a very annoying experience.

Improve the experience by updating the `<Lint/>` task so that when
`lint` reports file paths within `$(IntermediateOutputPath)`, the
paths are replaced with *meaningful* paths *outside* of
`$(IntermediateOutputPath)`.  Paths which, when edited, won't be
replaced on the next build.  Paths which can *actually* fix the
reported issues.

[0]: https://developer.android.com/studio/write/lint