[tests] add checks if long paths supported (#3543)
Context: http://build.azdo.io/2971289

Our `MaxPathTests` (a3009f7b) seem to be failing on a specific build
machine: DDMBLDW137.  I suspect this machine has long paths enabled.

I added a check that attempts to write a long path, and if it
succeeds we can `Assert.Ignore()` these tests appropriately.

A filename can only be 255 characters, so I added a constant for this
value so the various tests can use it.